pyproject.toml
[project]
name = "pokemon-damage-calculator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserde>=0.25.1",
]

[project.scripts]

[tool.pyright]
reportRedeclaration = false

[tool.pytest.ini_options]
log_cli_level = "INFO"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.13.1",
]
